move exception stuff to one file
